Refactoring event listeners and functions
Move event listeners from main.ts to eventListeners.ts, delete unnecessary (double) eventListener in displayHandler, change handling of displaying chat and messages (via Message class) and change name of displayChat function to displayMessages, add showForum when logged in, update docs.
