+++ docs/index.543b44e8.js
@@ -1544,7 +1544,7 @@ class Tr{constructor(e,t,n,i){this.eventType=e,this.eventRegistration=t,this.sna
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-class Nr{constructor(e,t,n,i){this._repo=e,this._path=t,this._queryParams=n,this._orderByCalled=i}get key(){return At(this._path)?null:Pt(this._path)}get ref(){return new Pr(this._repo,this._path)}get _queryIdentifier(){const e=Rn(this._queryParams),t=Qe(e);return"{}"===t?"default":t}get _queryObject(){return Rn(this._queryParams)}isEqual(e){if(!((e=R(e))instanceof Nr))return!1;const t=this._repo===e._repo,n=Ot(this._path,e._path),i=this._queryIdentifier===e._queryIdentifier;return t&&n&&i}toJSON(){return this.toString()}toString(){return this._repo.toString()+function(e){let t="";for(let n=e.pieceNum_;n<e.pieces_.length;n++)""!==e.pieces_[n]&&(t+="/"+encodeURIComponent(String(e.pieces_[n])));return t||"/"}(this._path)}}class Pr extends Nr{constructor(e,t){super(e,t,new Pn,!1)}get parent(){const e=Rt(this._path);return null===e?null:new Pr(this._repo,e)}get root(){let e=this;for(;null!==e.parent;)e=e.parent;return e}}class xr{constructor(e,t,n){this._node=e,this.ref=t,this._index=n}get priority(){return this._node.getPriority().val()}get key(){return this.ref.key}get size(){return this._node.numChildren()}child(e){const t=new Et(e),n=Dr(this.ref,e);return new xr(this._node.getChild(t),n,an)}exists(){return!this._node.isEmpty()}exportVal(){return this._node.val(!0)}forEach(e){if(this._node.isLeafNode())return!1;return!!this._node.forEachChild(this._index,((t,n)=>e(new xr(n,Dr(this.ref,t),an))))}hasChild(e){const t=new Et(e);return!this._node.getChild(t).isEmpty()}hasChildren(){return!this._node.isLeafNode()&&!this._node.isEmpty()}toJSON(){return this.exportVal()}val(){return this._node.val()}}function Rr(e,t){return(e=R(e))._checkNotDeleted("ref"),void 0!==t?Dr(e._root,t):e._root}function Dr(e,t){var n,i,s,r;return null===St((e=R(e))._path)?(n="child",i="path",r=!1,(s=t)&&(s=s.replace(/^\/*\.info(\/|$)/,"/")),$s(n,i,s,r)):$s("child","path",t,!1),new Pr(e._repo,Dt(e._path,t))}function Ar(e,t){e=R(e),Ys("push",e._path),Bs("push",t,e._path,!0);const n=ir(e._repo),i=wn(n),s=Dr(e,i),r=Dr(e,i);let o;return o=null!=t?Fr(r,t).then((()=>r)):Promise.resolve(r),s.then=o.then.bind(o),s.catch=o.then.bind(o,void 0),s}function Mr(e){return Ys("remove",e._path),Fr(e,null)}function Fr(e,t){e=R(e),Ys("set",e._path),Bs("set",t,e._path,!1);const n=new d;return hr(e._repo,e._path,t,null,n.wrapCallback((()=>{}))),n.promise}function Or(e,t){Vs("update",t,e._path,!1);const n=new d;return function(e,t,n,i){dr(e,"update",{path:t.toString(),value:n});let s=!0;const r=sr(e),o={};if(Je(n,((n,i)=>{s=!1,o[n]=Ss(Dt(t,n),mn(i),e.serverSyncTree_,r)})),s)He("update() called with empty data.  Don't do anything."),pr(0,i,"ok",void 0);else{const s=lr(e),r=rs(e.serverSyncTree_,t,o,s);Qs(e.eventQueue_,r),e.server_.merge(t.toString(),n,((n,r)=>{const o="ok"===n;o||Ve("update at "+t+" failed: "+n);const a=os(e.serverSyncTree_,s,!o),l=a.length>0?gr(e,t):t;Js(e.eventQueue_,l,a),pr(0,i,n,r)})),Je(n,(n=>{const i=wr(e,Dt(t,n));gr(e,i)})),Js(e.eventQueue_,t,[])}}(e._repo,e._path,t,n.wrapCallback((()=>{}))),n.promise}class Lr{constructor(e){this.callbackContext=e}respondsTo(e){return"value"===e}createEvent(e,t){const n=t._queryParams.getIndex();return new Tr("value",this,new xr(e.snapshotNode,new Pr(t._repo,t._path),n))}getEventRunner(e){return"cancel"===e.getEventType()?()=>this.callbackContext.onCancel(e.error):()=>this.callbackContext.onValue(e.snapshot,null)}createCancelEvent(e,t){return this.callbackContext.hasCancelCallback?new Sr(this,e,t):null}matches(e){return e instanceof Lr&&(!e.callbackContext||!this.callbackContext||e.callbackContext.matches(this.callbackContext))}hasAnyCallback(){return null!==this.callbackContext}}class qr{constructor(e,t){this.eventType=e,this.callbackContext=t}respondsTo(e){let t="children_added"===e?"child_added":e;return t="children_removed"===t?"child_removed":t,this.eventType===t}createCancelEvent(e,t){return this.callbackContext.hasCancelCallback?new Sr(this,e,t):null}createEvent(e,t){i(null!=e.childName,"Child events should have a childName.");const n=Dr(new Pr(t._repo,t._path),e.childName),s=t._queryParams.getIndex();return new Tr(e.type,this,new xr(e.snapshotNode,n,s),e.prevName)}getEventRunner(e){return"cancel"===e.getEventType()?()=>this.callbackContext.onCancel(e.error):()=>this.callbackContext.onValue(e.snapshot,e.prevName)}matches(e){return e instanceof qr&&(this.eventType===e.eventType&&(!this.callbackContext||!e.callbackContext||this.callbackContext.matches(e.callbackContext)))}hasAnyCallback(){return!!this.callbackContext}}function Ur(e,t,n,i,s){let r;if("object"==typeof i&&(r=void 0,s=i),"function"==typeof i&&(r=i),s&&s.onlyOnce){const t=n,i=(n,i)=>{cr(e._repo,e,a),t(n,i)};i.userCallback=n.userCallback,i.context=n.context,n=i}const o=new kr(n,r||void 0),a="value"===t?new Lr(o):new qr(t,o);return function(e,t,n){let i;i=".info"===St(t._path)?hs(e.infoSyncTree_,t,n):hs(e.serverSyncTree_,t,n),Xs(e.eventQueue_,t._path,i)}(e._repo,e,a),()=>cr(e._repo,e,a)}function Wr(e,t,n,i){return Ur(e,"value",t,n,i)}!function(e){i(!Bi,"__referenceConstructor has already been defined"),Bi=e}(Pr),function(e){i(!zi,"__referenceConstructor has already been defined"),zi=e}(Pr);
+class Nr{constructor(e,t,n,i){this._repo=e,this._path=t,this._queryParams=n,this._orderByCalled=i}get key(){return At(this._path)?null:Pt(this._path)}get ref(){return new Pr(this._repo,this._path)}get _queryIdentifier(){const e=Rn(this._queryParams),t=Qe(e);return"{}"===t?"default":t}get _queryObject(){return Rn(this._queryParams)}isEqual(e){if(!((e=R(e))instanceof Nr))return!1;const t=this._repo===e._repo,n=Ot(this._path,e._path),i=this._queryIdentifier===e._queryIdentifier;return t&&n&&i}toJSON(){return this.toString()}toString(){return this._repo.toString()+function(e){let t="";for(let n=e.pieceNum_;n<e.pieces_.length;n++)""!==e.pieces_[n]&&(t+="/"+encodeURIComponent(String(e.pieces_[n])));return t||"/"}(this._path)}}class Pr extends Nr{constructor(e,t){super(e,t,new Pn,!1)}get parent(){const e=Rt(this._path);return null===e?null:new Pr(this._repo,e)}get root(){let e=this;for(;null!==e.parent;)e=e.parent;return e}}class xr{constructor(e,t,n){this._node=e,this.ref=t,this._index=n}get priority(){return this._node.getPriority().val()}get key(){return this.ref.key}get size(){return this._node.numChildren()}child(e){const t=new Et(e),n=Dr(this.ref,e);return new xr(this._node.getChild(t),n,an)}exists(){return!this._node.isEmpty()}exportVal(){return this._node.val(!0)}forEach(e){if(this._node.isLeafNode())return!1;return!!this._node.forEachChild(this._index,((t,n)=>e(new xr(n,Dr(this.ref,t),an))))}hasChild(e){const t=new Et(e);return!this._node.getChild(t).isEmpty()}hasChildren(){return!this._node.isLeafNode()&&!this._node.isEmpty()}toJSON(){return this.exportVal()}val(){return this._node.val()}}function Rr(e,t){return(e=R(e))._checkNotDeleted("ref"),void 0!==t?Dr(e._root,t):e._root}function Dr(e,t){var n,i,s,r;return null===St((e=R(e))._path)?(n="child",i="path",r=!1,(s=t)&&(s=s.replace(/^\/*\.info(\/|$)/,"/")),$s(n,i,s,r)):$s("child","path",t,!1),new Pr(e._repo,Dt(e._path,t))}function Ar(e,t){e=R(e),Ys("push",e._path),Bs("push",t,e._path,!0);const n=ir(e._repo),i=wn(n),s=Dr(e,i),r=Dr(e,i);let o;return o=null!=t?Mr(r,t).then((()=>r)):Promise.resolve(r),s.then=o.then.bind(o),s.catch=o.then.bind(o,void 0),s}function Mr(e,t){e=R(e),Ys("set",e._path),Bs("set",t,e._path,!1);const n=new d;return hr(e._repo,e._path,t,null,n.wrapCallback((()=>{}))),n.promise}function Fr(e,t){Vs("update",t,e._path,!1);const n=new d;return function(e,t,n,i){dr(e,"update",{path:t.toString(),value:n});let s=!0;const r=sr(e),o={};if(Je(n,((n,i)=>{s=!1,o[n]=Ss(Dt(t,n),mn(i),e.serverSyncTree_,r)})),s)He("update() called with empty data.  Don't do anything."),pr(0,i,"ok",void 0);else{const s=lr(e),r=rs(e.serverSyncTree_,t,o,s);Qs(e.eventQueue_,r),e.server_.merge(t.toString(),n,((n,r)=>{const o="ok"===n;o||Ve("update at "+t+" failed: "+n);const a=os(e.serverSyncTree_,s,!o),l=a.length>0?gr(e,t):t;Js(e.eventQueue_,l,a),pr(0,i,n,r)})),Je(n,(n=>{const i=wr(e,Dt(t,n));gr(e,i)})),Js(e.eventQueue_,t,[])}}(e._repo,e._path,t,n.wrapCallback((()=>{}))),n.promise}class Or{constructor(e){this.callbackContext=e}respondsTo(e){return"value"===e}createEvent(e,t){const n=t._queryParams.getIndex();return new Tr("value",this,new xr(e.snapshotNode,new Pr(t._repo,t._path),n))}getEventRunner(e){return"cancel"===e.getEventType()?()=>this.callbackContext.onCancel(e.error):()=>this.callbackContext.onValue(e.snapshot,null)}createCancelEvent(e,t){return this.callbackContext.hasCancelCallback?new Sr(this,e,t):null}matches(e){return e instanceof Or&&(!e.callbackContext||!this.callbackContext||e.callbackContext.matches(this.callbackContext))}hasAnyCallback(){return null!==this.callbackContext}}class Lr{constructor(e,t){this.eventType=e,this.callbackContext=t}respondsTo(e){let t="children_added"===e?"child_added":e;return t="children_removed"===t?"child_removed":t,this.eventType===t}createCancelEvent(e,t){return this.callbackContext.hasCancelCallback?new Sr(this,e,t):null}createEvent(e,t){i(null!=e.childName,"Child events should have a childName.");const n=Dr(new Pr(t._repo,t._path),e.childName),s=t._queryParams.getIndex();return new Tr(e.type,this,new xr(e.snapshotNode,n,s),e.prevName)}getEventRunner(e){return"cancel"===e.getEventType()?()=>this.callbackContext.onCancel(e.error):()=>this.callbackContext.onValue(e.snapshot,e.prevName)}matches(e){return e instanceof Lr&&(this.eventType===e.eventType&&(!this.callbackContext||!e.callbackContext||this.callbackContext.matches(e.callbackContext)))}hasAnyCallback(){return!!this.callbackContext}}function qr(e,t,n,i,s){let r;if("object"==typeof i&&(r=void 0,s=i),"function"==typeof i&&(r=i),s&&s.onlyOnce){const t=n,i=(n,i)=>{cr(e._repo,e,a),t(n,i)};i.userCallback=n.userCallback,i.context=n.context,n=i}const o=new kr(n,r||void 0),a="value"===t?new Or(o):new Lr(t,o);return function(e,t,n){let i;i=".info"===St(t._path)?hs(e.infoSyncTree_,t,n):hs(e.serverSyncTree_,t,n),Xs(e.eventQueue_,t._path,i)}(e._repo,e,a),()=>cr(e._repo,e,a)}function Ur(e,t,n,i){return qr(e,"value",t,n,i)}!function(e){i(!Bi,"__referenceConstructor has already been defined"),Bi=e}(Pr),function(e){i(!zi,"__referenceConstructor has already been defined"),zi=e}(Pr);
 /**
  * @license
  * Copyright 2020 Google LLC
@@ -1561,7 +1561,7 @@ class Nr{constructor(e,t,n,i){this._repo=e,this._path=t,this._queryParams=n,this
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-const Hr={};let jr=!1;function Br(e,t,n,i,s){let r=i||e.options.databaseURL;void 0===r&&(e.options.projectId||ze("Can't determine Firebase Database URL. Be sure to include  a Project ID when calling firebase.initializeApp()."),He("Using default host for project ",e.options.projectId),r=`${e.options.projectId}-default-rtdb.firebaseio.com`);let o,a,l=Ir(r,s),h=l.repoInfo;void 0!==pe&&(a=pe.env.FIREBASE_DATABASE_EMULATOR_HOST),a?(o=!0,r=`http://${a}?ns=${h.namespace}`,l=Ir(r,s),h=l.repoInfo):o=!l.repoInfo.secure;const c=s&&o?new ot(ot.OWNER):new rt(e.name,e.options,t);Gs("Invalid Firebase Database URL",l),At(l.path)||ze("Database URL must point to the root of a Firebase Database (not including a child path).");const u=function(e,t,n,i){let s=Hr[t.name];s||(s={},Hr[t.name]=s);let r=s[e.toURLString()];r&&ze("Database initialized multiple times. Please make sure the format of the database URL matches with each database() call.");return r=new tr(e,jr,n,i),s[e.toURLString()]=r,r}(h,e,c,new st(e.name,n));return new zr(u,e)}class zr{constructor(e,t){this._repoInternal=e,this.app=t,this.type="database",this._instanceStarted=!1}get _repo(){return this._instanceStarted||(nr(this._repoInternal,this.app.options.appId,this.app.options.databaseAuthVariableOverride),this._instanceStarted=!0),this._repoInternal}get _root(){return this._rootInternal||(this._rootInternal=new Pr(this._repo,Tt())),this._rootInternal}_delete(){return null!==this._rootInternal&&(!function(e,t){const n=Hr[t];n&&n[e.key]===e||ze(`Database ${t}(${e.repoInfo_}) has already been deleted.`),ur(e),delete n[e.key]}(this._repo,this.app.name),this._repoInternal=null,this._rootInternal=null),Promise.resolve()}_checkNotDeleted(e){null===this._rootInternal&&ze("Cannot call "+e+" on a deleted database.")}}jt.prototype.simpleListen=function(e,t){this.sendRequest("q",{p:e},t)},jt.prototype.echo=function(e,t){this.sendRequest("echo",{d:e},t)};!
+const Wr={};let Hr=!1;function jr(e,t,n,i,s){let r=i||e.options.databaseURL;void 0===r&&(e.options.projectId||ze("Can't determine Firebase Database URL. Be sure to include  a Project ID when calling firebase.initializeApp()."),He("Using default host for project ",e.options.projectId),r=`${e.options.projectId}-default-rtdb.firebaseio.com`);let o,a,l=Ir(r,s),h=l.repoInfo;void 0!==pe&&(a=pe.env.FIREBASE_DATABASE_EMULATOR_HOST),a?(o=!0,r=`http://${a}?ns=${h.namespace}`,l=Ir(r,s),h=l.repoInfo):o=!l.repoInfo.secure;const c=s&&o?new ot(ot.OWNER):new rt(e.name,e.options,t);Gs("Invalid Firebase Database URL",l),At(l.path)||ze("Database URL must point to the root of a Firebase Database (not including a child path).");const u=function(e,t,n,i){let s=Wr[t.name];s||(s={},Wr[t.name]=s);let r=s[e.toURLString()];r&&ze("Database initialized multiple times. Please make sure the format of the database URL matches with each database() call.");return r=new tr(e,Hr,n,i),s[e.toURLString()]=r,r}(h,e,c,new st(e.name,n));return new Br(u,e)}class Br{constructor(e,t){this._repoInternal=e,this.app=t,this.type="database",this._instanceStarted=!1}get _repo(){return this._instanceStarted||(nr(this._repoInternal,this.app.options.appId,this.app.options.databaseAuthVariableOverride),this._instanceStarted=!0),this._repoInternal}get _root(){return this._rootInternal||(this._rootInternal=new Pr(this._repo,Tt())),this._rootInternal}_delete(){return null!==this._rootInternal&&(!function(e,t){const n=Wr[t];n&&n[e.key]===e||ze(`Database ${t}(${e.repoInfo_}) has already been deleted.`),ur(e),delete n[e.key]}(this._repo,this.app.name),this._repoInternal=null,this._rootInternal=null),Promise.resolve()}_checkNotDeleted(e){null===this._rootInternal&&ze("Cannot call "+e+" on a deleted database.")}}jt.prototype.simpleListen=function(e,t){this.sendRequest("q",{p:e},t)},jt.prototype.echo=function(e,t){this.sendRequest("echo",{d:e},t)};!
 /**
  * @license
  * Copyright 2021 Google LLC
@@ -1578,7 +1578,7 @@ const Hr={};let jr=!1;function Br(e,t,n,i,s){let r=i||e.options.databaseURL;void
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-function(e){Ne="9.6.10",Z(new L("database",((e,{instanceIdentifier:t})=>Br(e.getProvider("app").getImmediate(),e.getProvider("auth-internal"),e.getProvider("app-check-internal"),t)),"PUBLIC").setMultipleInstances(!0)),ie("@firebase/database","0.12.7",e),ie("@firebase/database","0.12.7","esm2017")}
+function(e){Ne="9.6.10",Z(new L("database",((e,{instanceIdentifier:t})=>jr(e.getProvider("app").getImmediate(),e.getProvider("auth-internal"),e.getProvider("app-check-internal"),t)),"PUBLIC").setMultipleInstances(!0)),ie("@firebase/database","0.12.7",e),ie("@firebase/database","0.12.7","esm2017")}
 /**
  * @license
  * Copyright 2020 Google LLC
@@ -1611,5 +1611,5 @@ function(e){Ne="9.6.10",Z(new L("database",((e,{instanceIdentifier:t})=>Br(e.get
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-ie("firebase","9.6.10","app");const Vr=document.querySelector("#travel"),$r=document.querySelector("#sport"),Yr=document.querySelector("#gaming"),Gr=document.querySelector("#travel-chat"),Kr=document.querySelector("#sport-chat"),Qr=document.querySelector("#gaming-chat");let Xr=Vr,Jr=Gr,Zr=!0,eo=!1,to=!1;const no=e=>{"travel-forum"==e?(Xr=Vr,Jr=Gr,Vr.style.display="flex",$r.style.display="none",Yr.style.display="none",Zr=!0,eo=!1,to=!1,_o()):"sport-forum"==e?(Xr=$r,Jr=Kr,Vr.style.display="none",$r.style.display="flex",Yr.style.display="none",Zr=!1,eo=!0,to=!1,_o()):"gaming-forum"==e&&(Xr=Yr,Jr=Qr,Vr.style.display="none",$r.style.display="none",Yr.style.display="flex",Zr=!1,eo=!1,to=!0,_o())},io=e=>{const t=document.querySelector(".user"),n=document.createElement("div"),i=document.createElement("div"),s=document.createElement("h5"),r=document.createElement("h5"),o=document.createElement("p"),a=document.createElement("button"),l=document.createElement("div");s.innerText=e.username,r.innerText=e.timestamp,o.innerText=e.message,a.innerText="X",a.className="delete-btn",n.className="user-div",i.className="mess-div",l.className="message-container",l.id=e.id,n.append(s),n.append(r),i.append(o),e.username==t.innerText&&e.userId==t.id&&(i.append(a),a.addEventListener("click",(()=>{console.log("delete button clicked"),console.log(e.id),console.log(Xr.id);const t=Rr(ro,"/messages/"+Xr.id+"/"+e.id);console.log(t),Mr(t)}))),l.append(n),l.append(i),Jr.prepend(l),Xr.append(Jr),a.addEventListener("click",(()=>{console.log("delete button clicked"),console.log(e.id),console.log(Xr.id);const t=Rr(ro,"/messages/"+Xr.id+"/"+e.id);console.log(t),Mr(t)}))};class so{constructor(e,t,n,i,s){this.id=e,this.username=t,this.message=n,this.timestamp=i,this.userId=s}clearChat(){document.querySelector(`#${this.id}`).remove()}}const ro=function(e=function(e="[DEFAULT]"){const t=Q.get(e);if(!t)throw te.create("no-app",{appName:e});return t}(),t){return ee(e,"database").getImmediate({identifier:t})}(function(e,t={}){if("object"!=typeof t){t={name:t}}const n=Object.assign({name:"[DEFAULT]",automaticDataCollectionEnabled:!1},t),i=n.name;if("string"!=typeof i||!i)throw te.create("bad-app-name",{appName:String(i)});const s=Q.get(i);if(s){if(S(e,s.options)&&S(n,s.config))return s;throw te.create("duplicate-app",{appName:i})}const r=new U(i);for(const e of X.values())r.addComponent(e);const o=new ne(e,n,r);return Q.set(i,o),o}({apiKey:"AIzaSyBSQ4EwRXTEF6Rs_3REQTWtcOoBu3HOKEQ",authDomain:"gritbook-social-media-app.firebaseapp.com",databaseURL:"https://gritbook-social-media-app-default-rtdb.europe-west1.firebasedatabase.app",projectId:"gritbook-social-media-app",storageBucket:"gritbook-social-media-app.appspot.com",messagingSenderId:"110931333305",appId:"1:110931333305:web:c540e52ab897acade50b68"})),oo=Rr(ro,"/messages/travel"),ao=Rr(ro,"messages/sport"),lo=Rr(ro,"messages/gaming"),ho=Rr(ro,"/users");let co,uo=[];const po=()=>{!0===Zr?co=oo:!0===eo?co=ao:!0===to&&(co=lo)},_o=()=>{po(),Wr(co,(e=>{const t=e.val();console.log(t);for(const e of uo)e.clearChat();uo=[];for(const e in t)uo.push(new so(e,t[e].username,t[e].message,t[e].timestamp,t[e].userId));console.log(uo);for(const e of uo)io(e)}))},fo=()=>{let e;return Wr(ho,(t=>{e=t.val(),console.log(e)})),e};fo();class go{constructor(e,t){this.username=e,this.password=t}getUsername(){return this.username}getPassword(){return this.password}}const mo=document.querySelectorAll(".forum-btn"),yo=document.querySelector(".user"),vo=document.querySelector("#message"),Co=document.querySelector("#add-message-btn");new class{constructor(){this.signInForm=document.querySelector("#login-form"),this.signUpForm=document.querySelector("#signup-form"),this.username=document.querySelector("#username"),this.password=document.querySelector("#password"),document.querySelector("#signin-btn").addEventListener("click",this.validator.bind(this)),document.querySelector("#btn-signup-ui").addEventListener("click",this.hideUI.bind(this))}validator(e){if(e.preventDefault(),console.log("validator invoked"),console.log(this.username.value),console.log(this.password.value),this.username.value.length>=4&&this.password.value.length>=4){const e=fo();if(e){for(const t in e)if(this.username.value==e[t].username&&this.password.value==e[t].password){console.log("login success");const n=document.querySelector(".user");return n.textContent=e[t].username,n.setAttribute("id",t),void this.clearForm()}console.log("login failed"),this.clearForm()}}else console.log("username/password length is less than 4 characters"),this.clearForm()}hideUI(e){e.preventDefault(),this.clearForm(),this.signInForm.style.display="none",this.signUpForm.style.display="flex"}clearForm(){this.signInForm.reset()}},new class{constructor(){this.usernameInField=document.querySelector("#new_username"),this.passwordInField=document.querySelector("#new_password"),this.repeatPasswordInField=document.querySelector("#repeat_password"),this.signUpform=document.querySelector("#signup-form"),this.signInForm=document.querySelector("#login-form"),document.querySelector("#signup-btn").addEventListener("click",this.validateForm.bind(this)),document.querySelector("#btn-display-signin").addEventListener("click",this.hideUI.bind(this))}validateForm(e){e.preventDefault();const t=fo();let n=[];for(const e in t)n.push(t[e].username);this.usernameInField.value.length<4||this.passwordInField.value.length<4||this.repeatPasswordInField.value.length<4?alert("username/password length is less than 4 characters"):n.includes(this.usernameInField.value)?alert("username already exists"):this.passwordInField.value==this.repeatPasswordInField.value?(alert("signup success"),this.initNewUser(new go(this.usernameInField.value,this.passwordInField.value))):alert("password does not match")}hideUI(e){e.preventDefault(),console.log("hideUI invoked"),this.signUpform.style.display="none",this.signInForm.style.display="flex"}initNewUser(e){console.log("initNewUser invoked");const t={username:e.getUsername(),password:e.getPassword()},n=Rr(ro,"/users"),i={};i[Ar(n).key]=t,Or(n,i),this.clearForm()}clearForm(){this.signUpform.reset()}},no("travel-forum"),mo.forEach((e=>{e.addEventListener("click",(e=>{const t=e.target.id;no(t)}))})),Co.addEventListener("click",(e=>{e.preventDefault(),po();const t=Date.now(),n=new Date(t),i=e=>e<10?"0"+e:e,s=i(n.getDate())+"-"+i(n.getMonth()+1)+"-"+n.getFullYear()+", "+i(n.getHours())+":"+i(n.getMinutes());if(""==vo.value)alert("Enter text to post!");else{const e={message:vo.value,username:yo.innerText,timestamp:s,userId:yo.id};vo.value="";const t={};t[Ar(co).key]=e,Or(co,t)}}));
-//# sourceMappingURL=index.365cb23a.js.map
+ie("firebase","9.6.10","app");class zr{constructor(e,t,n,i,s){this.id=e,this.username=t,this.message=n,this.timestamp=i,this.userId=s}clearChat(){document.querySelector(`#${this.id}`).remove()}displayChat(){_o(this)}}const Vr=function(e=function(e="[DEFAULT]"){const t=Q.get(e);if(!t)throw te.create("no-app",{appName:e});return t}(),t){return ee(e,"database").getImmediate({identifier:t})}(function(e,t={}){if("object"!=typeof t){t={name:t}}const n=Object.assign({name:"[DEFAULT]",automaticDataCollectionEnabled:!1},t),i=n.name;if("string"!=typeof i||!i)throw te.create("bad-app-name",{appName:String(i)});const s=Q.get(i);if(s){if(S(e,s.options)&&S(n,s.config))return s;throw te.create("duplicate-app",{appName:i})}const r=new U(i);for(const e of X.values())r.addComponent(e);const o=new ne(e,n,r);return Q.set(i,o),o}({apiKey:"AIzaSyBSQ4EwRXTEF6Rs_3REQTWtcOoBu3HOKEQ",authDomain:"gritbook-social-media-app.firebaseapp.com",databaseURL:"https://gritbook-social-media-app-default-rtdb.europe-west1.firebasedatabase.app",projectId:"gritbook-social-media-app",storageBucket:"gritbook-social-media-app.appspot.com",messagingSenderId:"110931333305",appId:"1:110931333305:web:c540e52ab897acade50b68"})),$r=Rr(Vr,"/messages/travel"),Yr=Rr(Vr,"messages/sport"),Gr=Rr(Vr,"messages/gaming"),Kr=Rr(Vr,"/users");let Qr,Xr=[];const Jr=()=>{!0===ho?Qr=$r:!0===co?Qr=Yr:!0===uo&&(Qr=Gr)},Zr=()=>{Jr(),Ur(Qr,(e=>{const t=e.val();console.log(t);for(const e of Xr)e.clearChat();Xr=[];for(const e in t)Xr.push(new zr(e,t[e].username,t[e].message,t[e].timestamp,t[e].userId));console.log(Xr);for(const e of Xr)e.displayChat()}))},eo=()=>{let e;return Ur(Kr,(t=>{e=t.val(),console.log(e)})),e};eo();const to=document.querySelector("#travel"),no=document.querySelector("#sport"),io=document.querySelector("#gaming"),so=document.querySelector("#travel-chat"),ro=document.querySelector("#sport-chat"),oo=document.querySelector("#gaming-chat");let ao=to,lo=so,ho=!0,co=!1,uo=!1;const po=e=>{"travel-forum"==e?(ao=to,lo=so,to.style.display="flex",no.style.display="none",io.style.display="none",ho=!0,co=!1,uo=!1,Zr()):"sport-forum"==e?(ao=no,lo=ro,to.style.display="none",no.style.display="flex",io.style.display="none",ho=!1,co=!0,uo=!1,Zr()):"gaming-forum"==e&&(ao=io,lo=oo,to.style.display="none",no.style.display="none",io.style.display="flex",ho=!1,co=!1,uo=!0,Zr())},_o=e=>{const t=document.querySelector(".user"),n=document.createElement("div"),i=document.createElement("div"),s=document.createElement("h5"),r=document.createElement("h5"),o=document.createElement("p"),a=document.createElement("button"),l=document.createElement("div");s.innerText=e.username,r.innerText=e.timestamp,o.innerText=e.message,a.innerText="X",a.className="delete-btn",n.className="user-div",i.className="mess-div",l.className="message-container",l.id=e.id,n.append(s),n.append(r),i.append(o),e.username==t.innerText&&e.userId==t.id&&(i.append(a),a.addEventListener("click",(()=>{!function(e){Ys("remove",e._path),Mr(e,null)}(Rr(Vr,"/messages/"+ao.id+"/"+e.id))}))),l.append(n),l.append(i),lo.prepend(l),ao.append(lo)};class fo{constructor(e,t){this.username=e,this.password=t}getUsername(){return this.username}getPassword(){return this.password}}const go=document.querySelector(".user"),mo=document.querySelector("#message"),yo=e=>{e.preventDefault();const t=e.target.id;po(t)},vo=document.querySelectorAll(".forum-btn"),Co=document.querySelector("#add-message-btn");new class{constructor(){this.signInForm=document.querySelector("#login-form"),this.signUpForm=document.querySelector("#signup-form"),this.username=document.querySelector("#username"),this.password=document.querySelector("#password"),document.querySelector("#signin-btn").addEventListener("click",this.validator.bind(this)),document.querySelector("#btn-signup-ui").addEventListener("click",this.hideUI.bind(this))}validator(e){if(e.preventDefault(),console.log("validator invoked"),console.log(this.username.value),console.log(this.password.value),this.username.value.length>=4&&this.password.value.length>=4){const e=eo();if(e){for(const t in e)if(this.username.value==e[t].username&&this.password.value==e[t].password){console.log("login success");const n=document.querySelector(".user");return n.textContent=e[t].username,n.setAttribute("id",t),this.clearForm(),void po("travel-forum")}console.log("login failed"),this.clearForm()}}else console.log("username/password length is less than 4 characters"),this.clearForm()}hideUI(e){e.preventDefault(),this.clearForm(),this.signInForm.style.display="none",this.signUpForm.style.display="flex"}clearForm(){this.signInForm.reset()}},new class{constructor(){this.usernameInField=document.querySelector("#new_username"),this.passwordInField=document.querySelector("#new_password"),this.repeatPasswordInField=document.querySelector("#repeat_password"),this.signUpform=document.querySelector("#signup-form"),this.signInForm=document.querySelector("#login-form"),document.querySelector("#signup-btn").addEventListener("click",this.validateForm.bind(this)),document.querySelector("#btn-display-signin").addEventListener("click",this.hideUI.bind(this))}validateForm(e){e.preventDefault();const t=eo();let n=[];for(const e in t)n.push(t[e].username);this.usernameInField.value.length<4||this.passwordInField.value.length<4||this.repeatPasswordInField.value.length<4?alert("username/password length is less than 4 characters"):n.includes(this.usernameInField.value)?alert("username already exists"):this.passwordInField.value==this.repeatPasswordInField.value?(alert("signup success"),this.initNewUser(new fo(this.usernameInField.value,this.passwordInField.value))):alert("password does not match")}hideUI(e){e.preventDefault(),console.log("hideUI invoked"),this.signUpform.style.display="none",this.signInForm.style.display="flex"}initNewUser(e){console.log("initNewUser invoked");const t={username:e.getUsername(),password:e.getPassword()},n=Rr(Vr,"/users"),i={};i[Ar(n).key]=t,Fr(n,i),this.clearForm()}clearForm(){this.signUpform.reset()}},po("travel-forum"),vo.forEach((e=>{e.addEventListener("click",yo)})),Co.addEventListener("click",(e=>{e.preventDefault(),Jr();const t=Date.now(),n=new Date(t),i=e=>e<10?"0"+e:e,s=i(n.getDate())+"-"+i(n.getMonth()+1)+"-"+n.getFullYear()+", "+i(n.getHours())+":"+i(n.getMinutes());if(""==mo.value)alert("Enter text to post!");else{const e={message:mo.value,username:go.innerText,timestamp:s,userId:go.id},t={};t[Ar(Qr).key]=e,Fr(Qr,t),mo.value=""}}));
+//# sourceMappingURL=index.543b44e8.js.map
