+++ docs/index.html
@@ -1,1 +1,1 @@
-<!DOCTYPE html><html lang="en"><head><meta charset="utf-8"><meta http-equiv="X-UA-Compatible" content="IE=edge"><meta name="viewport" content="width=device-width, initial-scale=1.0"><script src="https://kit.fontawesome.com/3318035e44.js"></script><script type="module" src="index.365cb23a.js"></script><link rel="stylesheet" href="index.d4c6579b.css"><link rel="preconnect" href="https://fonts.googleapis.com"><link rel="preconnect" href="https://fonts.gstatic.com" crossorigin><link href="https://fonts.googleapis.com/css2?family=Inknut+Antiqua&display=swap" rel="stylesheet"><link rel="icon" href="conversation_16.6bac1f21.png" type="image/png" sizes="16x16"><title>Gritbook</title></head><body> <main>  <div id="header-data"> <div class="profile"> <i class="fa-solid fa-user-astronaut"></i> <h2 class="user">name</h2> </div>  <div class="logo"> <h1 id="home">Gritbook</h1> <img src="conversation_128.d2f83f14.png" alt="Gritbook logo" width="96" height="96"> </div> <button type="submit" class="buttons" id="logout-btn">Logout</button> </div>   <form class="col form-wrapper" id="login-form"> <div class="row"> <h3 class="account-title">Welcome back</h3> </div> <div class="row"> <input type="text" placeholder="Username" id="username" required> <input type="password" placeholder="Password" id="password" required> </div> <div class="col"> <button class="buttons" id="signin-btn">Sign In</button> <h5 class="account-title">Don't have an account?</h5> <button class="form-ui" id="btn-signup-ui" formnovalidate> Sign up </button> </div> </form>    <form class="col form-wrapper" id="signup-form" style="display:none"> <div class="row"> <h3 class="account-title">Create account</h3> </div> <div class="row"> <input type="text" placeholder="Username" id="new_username" required> <input type="password" placeholder="Password" id="new_password" required> <input type="password" placeholder="Repeat Password" id="repeat_password" required> </div> <div class="col"> <button class="buttons" id="signup-btn">Sign Up</button> <h5 class="account-title">Already have an account?</h5> <button class="form-ui" id="btn-display-signin" formnovalidate> Sign in </button> </div> </form>   <section id="forums-section">  <div id="sidebar"> <a href="#home"><i class="fa-house fa-solid"></i> Home</a> <a href="#travel" class="forum-btn" id="travel-forum"><i class="fa-plane fa-solid"></i> Travel</a> <a href="#sport" class="forum-btn" id="sport-forum"><i class="fa-basketball fa-solid"></i> Sport</a> <a href="#gaming" class="forum-btn" id="gaming-forum"><i class="fa-gamepad fa-solid"></i> Gaming</a> <a href="#profiles" class="forum-btn" id="users-profiles"><i class="fa-solid fa-user-secret"></i> Users</a> </div>   <div id="topics-section">  <form id="message-form"> <input type="text" placeholder="Enter message" id="message" required> <button type="submit" class="buttons" id="add-message-btn"> Enter </button> </form>   <div class="topics" id="travel"> <h2>Travel</h2> <div class="chat-container" id="travel-chat"></div> </div> <div class="topics" id="sport"> <h2>Sport</h2> <div class="chat-container" id="sport-chat"></div> </div> <div class="topics" id="gaming"> <h2>Gaming</h2> <div class="chat-container" id="gaming-chat"></div> </div>  </div>  </section>  </main>  <footer> <div class="footer"> <p class="copyright">Gritbook Ltd.</p> <p class="copyright">Copyright &copy; 2022 | Othman, Arbie, Gosia</p> <span id="chevron" title="To the top"> <a href="#"><i class="fa-chevron-circle-up fas"></i></a> </span> </div> </footer>  </body></html>
+<!DOCTYPE html><html lang="en"><head><meta charset="utf-8"><meta http-equiv="X-UA-Compatible" content="IE=edge"><meta name="viewport" content="width=device-width, initial-scale=1.0"><script src="https://kit.fontawesome.com/3318035e44.js"></script><script type="module" src="index.543b44e8.js"></script><link rel="stylesheet" href="index.d4c6579b.css"><link rel="preconnect" href="https://fonts.googleapis.com"><link rel="preconnect" href="https://fonts.gstatic.com" crossorigin><link href="https://fonts.googleapis.com/css2?family=Inknut+Antiqua&display=swap" rel="stylesheet"><link rel="icon" href="conversation_16.6bac1f21.png" type="image/png" sizes="16x16"><title>Gritbook</title></head><body> <main>  <div id="header-data"> <div class="profile"> <i class="fa-solid fa-user-astronaut"></i> <h2 class="user">name</h2> </div>  <div class="logo"> <h1 id="home">Gritbook</h1> <img src="conversation_128.d2f83f14.png" alt="Gritbook logo" width="96" height="96"> </div> <button type="submit" class="buttons" id="logout-btn">Logout</button> </div>   <form class="col form-wrapper" id="login-form"> <div class="row"> <h3 class="account-title">Welcome back</h3> </div> <div class="row"> <input type="text" placeholder="Username" id="username" required> <input type="password" placeholder="Password" id="password" required> </div> <div class="col"> <button class="buttons" id="signin-btn">Sign In</button> <h5 class="account-title">Don't have an account?</h5> <button class="form-ui" id="btn-signup-ui" formnovalidate> Sign up </button> </div> </form>    <form class="col form-wrapper" id="signup-form" style="display:none"> <div class="row"> <h3 class="account-title">Create account</h3> </div> <div class="row"> <input type="text" placeholder="Username" id="new_username" required> <input type="password" placeholder="Password" id="new_password" required> <input type="password" placeholder="Repeat Password" id="repeat_password" required> </div> <div class="col"> <button class="buttons" id="signup-btn">Sign Up</button> <h5 class="account-title">Already have an account?</h5> <button class="form-ui" id="btn-display-signin" formnovalidate> Sign in </button> </div> </form>   <section id="forums-section">  <div id="sidebar"> <a href="#home"><i class="fa-house fa-solid"></i> Home</a> <a href="#travel" class="forum-btn" id="travel-forum"><i class="fa-plane fa-solid"></i> Travel</a> <a href="#sport" class="forum-btn" id="sport-forum"><i class="fa-basketball fa-solid"></i> Sport</a> <a href="#gaming" class="forum-btn" id="gaming-forum"><i class="fa-gamepad fa-solid"></i> Gaming</a> <a href="#profiles" class="forum-btn" id="users-profiles"><i class="fa-solid fa-user-secret"></i> Users</a> </div>   <div id="topics-section">  <form id="message-form"> <input type="text" placeholder="Enter message" id="message" required> <button type="submit" class="buttons" id="add-message-btn"> Enter </button> </form>   <div class="topics" id="travel"> <h2>Travel</h2> <div class="chat-container" id="travel-chat"></div> </div> <div class="topics" id="sport"> <h2>Sport</h2> <div class="chat-container" id="sport-chat"></div> </div> <div class="topics" id="gaming"> <h2>Gaming</h2> <div class="chat-container" id="gaming-chat"></div> </div>  </div>  </section>  </main>  <footer> <div class="footer"> <p class="copyright">Gritbook Ltd.</p> <p class="copyright">Copyright &copy; 2022 | Othman, Arbie, Gosia</p> <span id="chevron" title="To the top"> <a href="#"><i class="fa-chevron-circle-up fas"></i></a> </span> </div> </footer>  </body></html>
